Move tests to mock/multimap and fix CI
